requirements.txt
# Web框架
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
websockets>=11.0.3
python-multipart>=0.0.6

# RAG框架
langchain>=0.1.0
langgraph>=0.0.15
langchain-community>=0.0.10
langchain-openai>=0.0.5

# PDF处理
pypdf>=3.15.1
pdfplumber>=0.10.2
pytesseract>=0.3.10
pillow>=10.0.0

# 机器学习
sentence-transformers>=2.2.2  # 启用用于语义分块
transformers>=4.30.0
torch>=2.0.0
openai>=1.3.0

# 向量数据库
chromadb>=0.4.18
faiss-cpu>=1.7.4

# 检索
rank-bm25>=0.2.2

# 数据库
pymysql>=1.1.0
redis>=5.0.0
sqlalchemy>=2.0.0
alembic>=1.12.0

# 工具库
python-dotenv>=1.0.0
pydantic>=2.0.0
numpy>=1.24.0
pandas>=2.0.0

# 开发工具
pytest>=7.4.3
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0